# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Rezervo API", "version": "0.1.0", "description": "Tenant-isolated booking API. Protected business calls require the authenticated cookie and x-business-id header." },
  "servers": [{ "url": "http://localhost:4000/api" }],
  "paths": {
    "/auth/register": { "post": { "summary": "Create an account", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterInput" } } } }, "responses": { "201": { "description": "Created" } } } },
    "/auth/login": { "post": { "summary": "Start secure session", "responses": { "200": { "description": "Authenticated" }, "401": { "$ref": "#/components/responses/ApiError" } } } },
...
  },
  "components": {
...
    "schemas": {
      "RegisterInput": { "type": "object", "required": ["firstName", "lastName", "email", "password"], "properties": { "firstName": { "type": "string" }, "lastName": { "type": "string" }, "email": { "type": "string", "format": "email" }, "password": { "type": "string", "minLength": 12 } } },
...
      "ApiError": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "error": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" } } } } }
    },
    "responses": { "ApiError": { "description": "Application error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } } }
  }
}